refactor: relocate capability registry and extension runtime to the shell
